B6 Wrapup of Changes (#149)
* Runaway event killer!

* Added ability to play sound effects that are not tied to a map.

* Image based hp/mp/casting bars. Also added shields to the entity hud.

* Quick change so fading only happens when you're actually loading into game.

* Small tweak to made shields appear a little cleaner

* Added 'Folders' to most game objects. Added sorting/searching/etc to AnimationEditor.

* Class Editor Sorting/Searching/Folders

* Common Events - Copy/Pasting, Sorting, Folders, and more!

* Crafting Tables - Sorting/Searching/Folders

* Crafts Editor - Sorting/Searching/Folders

* Item Editor - Sorting/Searching/Folders

* Npc Editor - Sorting/Searching/Folders

* Projectile Editor - Sorting/Searching/Folders

* Resource Editor - Sorting/Searching/Folders

* Shop Editor - Sorting/Searching/Folders

* Spell Editor - Sorting/Searching/Folders

* Quest Editor - Sorting/Searching/Folders

* Variable Editor - Sorting/Searching/Folders

* Added separate client font option for ui vs ingame.

* Fix credits text from auto expanding. We want to use what is loaded via the json ui files.

* Removed the automated resource grabber from the client. Resources will be distributed with the binaries from now on.

* Tweaking logging so we don't spam general user consoles but still appropriately log for debugging purporses.

* Fixed migrations so they work with Sqlite and Mysql

* Fixing variable value parsing.

* Added warning when migrations will occur for people to backup their games else risk losing data.

* Fixed positioning of large player/npc sprites.

* Fixed r1659

https://www.ascensiongamedev.com/bug_tracker/intersect/monster-loot-no-longer-resized-to-32x32-r1659

* Update Intersect.Client.Framework/Gwen/Control/Label.cs


* Cleanup, and fixing Panda's suggestions for PR149

Former-commit-id: dd42c43f03628296c3bc006f55b858d1587ac400

+++ Intersect.Server/Classes/Admin/Actions/ActionProcessing.cs
@@ -206,7 +206,7 @@ namespace Intersect.Server.Classes.Admin.Actions
             var target = Player.FindOnline(action.Name);
             if (target != null)
             {
-                player.Warp(target.MapId, target.X, target.Y);
+                player.Warp(target.MapId, (byte)target.X, (byte)target.Y);
                 PacketSender.SendChatMsg(client, Strings.Player.warpedto.ToString(target.Name));
                 PacketSender.SendChatMsg(target.Client, Strings.Player.warpedtoyou.ToString(player.Name));
             }
@@ -225,7 +225,7 @@ namespace Intersect.Server.Classes.Admin.Actions
         //WarpToMapAction
         public static void ProcessAction(Client client, Player player, WarpToMapAction action)
         {
-            player.Warp(action.MapId, player.X, player.Y);
+            player.Warp(action.MapId, (byte)player.X, (byte)player.Y);
         }
 
         //WarpToMeAction
@@ -234,7 +234,7 @@ namespace Intersect.Server.Classes.Admin.Actions
             var target = Player.FindOnline(action.Name);
             if (target != null)
             {
-                target.Warp(player.MapId, player.X, player.Y);
+                target.Warp(player.MapId, (byte)player.X, (byte)player.Y);
                 PacketSender.SendChatMsg(client, Strings.Player.haswarpedto.ToString(target.Name), player.Name);
                 PacketSender.SendChatMsg(target.Client, Strings.Player.beenwarpedto.ToString(player.Name), player.Name);
             }
